external services added for the user, question service.

+++ User profile Service/User-profile-Service/src/main/java/com/interflow/userservice/entity/User.java
@@ -54,9 +54,9 @@ public class User {
 
     private List<UUID> answeredQuestionsIds;
 
-    private AtomicInteger totalUpVotes;
+    private long totalUpVotes;
 
-    private AtomicInteger profileVisits;
+    private long profileVisits;
 
     private List<UUID> savedQuestions;
